# lib/api-zod/src/generated/types/committeeMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommitteeMember {
  id: number;
  nameEn: string;
  nameTe: string;
  nameHi: string;
  roleEn: string;
  roleTe: string;
  roleHi: string;
  groupKey: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}