# lib/api-zod/src/generated/types/committeeMemberUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommitteeMemberUpdate {
  /** @minLength 1 */
  nameEn?: string;
  nameTe?: string;
  nameHi?: string;
  /** @minLength 1 */
  roleEn?: string;
  roleTe?: string;
  roleHi?: string;
  groupKey?: string;
  sortOrder?: number;
}