# lib/api-zod/src/generated/types/galleryEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryEvent {
  id: number;
  titleEn: string;
  titleTe: string;
  titleHi: string;
  descriptionEn: string;
  descriptionTe: string;
  descriptionHi: string;
  eventDate: string;
  category: string;
  photoUrls: string[];
  createdAt: string;
  updatedAt: string;
}