# lib/api-zod/src/generated/types/galleryEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryEventInput {
  /** @minLength 1 */
  titleEn: string;
  titleTe?: string;
  titleHi?: string;
  descriptionEn?: string;
  descriptionTe?: string;
  descriptionHi?: string;
  /** @minLength 1 */
  eventDate: string;
  /** @minLength 1 */
  category: string;
  photoUrls: string[];
}