# lib/api-zod/src/generated/types/galleryEventUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryEventUpdate {
  /** @minLength 1 */
  titleEn?: string;
  titleTe?: string;
  titleHi?: string;
  descriptionEn?: string;
  descriptionTe?: string;
  descriptionHi?: string;
  eventDate?: string;
  category?: string;
  photoUrls?: string[];
}